srcs/requirements/bonus/typeform-back/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==4.2.19
django-filter==24.3
djangorestframework==3.15.2
importlib-metadata==8.6.1
Markdown==3.7
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
typing-extensions==4.12.2
zipp==3.21.0
